add exit

+++ class_client.py
@@ -46,6 +46,7 @@ class Sock():
         return buf
 
 
+
     def Recv_Data(self):
         data = 1
         while data != '':
@@ -61,12 +62,17 @@ class Sock():
                 break
         self.sock.close()
 
+    def Exit(self,send,recv):
+        while True:
+            if not(send.is_alive() or recv.is_alive()):
+                self.sock.close()
+                # When everything done, release the capture
+                self.capture.release()
+                cv2.destroyAllWindows()
+
     def Main(self):
 
         SelfProc = multiprocessing.Process(target=self.Get_Self_Img()).run()
         RecvProc = multiprocessing.Process(target=self.Recv_Data()).run()
 
-        self.sock.close()
-        # When everything done, release the capture
-        self.capture.release()
-        cv2.destroyAllWindows()
+        self.Exit(SelfProc,RecvProc)
